requirements.txt
astroid==1.6.1
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
docopt==0.6.2
flake8==3.5.0
Flask==0.10.1
Flask-Testing==0.4.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.3.1
locustio==0.8.1
MarkupSafe==0.23
mccabe==0.6.1
msgpack-python==0.5.4
nose==1.3.7
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.2
pytest==3.4.0
python-owasp-zap-v2.4==0.0.14
pyzmq==17.0.0
requests==2.18.4
selenium==3.9.0
six==1.11.0
urllib3==1.22
Werkzeug==0.11.9
wrapt==1.10.11
